fix: positioning styles
